+++ spec/lnrecall-view-spec.js
@@ -1,9 +1,0 @@
-'use babel';
-
-import LnrecallView from '../lib/lnrecall-view';
-
-describe('LnrecallView', () => {
-  it('has one valid test', () => {
-    expect('life').toBe('easy');
-  });
-});