+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 0.0.3
+* Updated spec
+* Cleaned up implementations
+
 ## 0.0.1 - Mark's and Recalls
 * Added Mark Line and Recall Line functionality.
 * Added Keybinds
 * Added Context Menu Hooks
-* Added status-bar-left links 
+* Added status-bar-left links